+++ libs/prime-ng-ext/shared/src/index.ts
@@ -1,3 +1,5 @@
+export * from './lib/card.interface';
+export * from './lib/images-carousel.interface';
 export * from './lib/simple-image.interface';
 export * from './lib/simple-table.interface';
 export * from './lib/simple-text.interface';