+++ libs/prime-ng-playground/shared/generate-json-schemas.ts
@@ -5,6 +5,8 @@ import { ZodType } from 'zod';
 import { zodToJsonSchema } from 'zod-to-json-schema';
 
 import {
+  ZAppEditCardUIESchema,
+  ZAppEditImagesCarouselUIESchema,
   ZAppEditSimpleGridUIESchema,
   ZAppEditSimpleImageUIESchema,
   ZAppEditSimpleTableUIESchema,
@@ -28,4 +30,6 @@ generateJSONSchemaFromZodTypes([
   ZAppEditSimpleTextUIESchema,
   ZAppEditSimpleImageUIESchema,
   ZAppEditSimpleGridUIESchema,
+  ZAppEditImagesCarouselUIESchema,
+  ZAppEditCardUIESchema,
 ]);