+++ libs/prime-ng-playground/shared/src/lib/app-template.ts
@@ -1,6 +1,8 @@
 import { ZSimpleGridLayoutUIESchema } from '@dj-ui/common/shared';
 import type { RemoteResourceTemplate, UIElementTemplate } from '@dj-ui/core';
 import {
+  ZCardUIESchema,
+  ZImagesCarouselUIESchema,
   ZSimpleImageUIESchema,
   ZSimpleTableUIESchema,
   ZSimpleTextUIESchema,
@@ -64,6 +66,12 @@ export const ZAppEditSimpleTextUIESchema = createAppUIEEditTemplate(ZSimpleTextU
 export const ZAppEditSimpleImageUIESchema = createAppUIEEditTemplate(
   ZSimpleImageUIESchema
 ).describe('AppEditPrimeNgSimpleImageUIESchema');
+export const ZAppEditImagesCarouselUIESchema = createAppUIEEditTemplate(
+  ZImagesCarouselUIESchema
+).describe('AppEditPrimeNgImagesCarouselUIESchema');
 export const ZAppEditSimpleGridUIESchema = createAppUIEEditTemplate(
   ZSimpleGridLayoutUIESchema
 ).describe('AppEditSimpleGridUIESchema');
+export const ZAppEditCardUIESchema = createAppUIEEditTemplate(ZCardUIESchema).describe(
+  'AppEditPrimeNgCardUIESchema'
+);